rtph264depay: fix an error when the total number of SPS was 32
The mask 0x1f would filter out any number greater than 31.
So in the codec data, we would write 0 as the number of SPS's when the 32nd SPS (with id 31) arrived.
Here we make sure that the array is no longer than 32, so we can write
in the codec_data the total number of SPS without any mask.